Added Google Places API functionality
Add navigator.geolocation to find the user's current location. However, this only works when viewing your file in the browser (i.e. file:///Volumes/Mactinosh etc...) instead of localhost. We may need to see if we can get the user's location using the Google Maps API instead, as Typekit will not work.

Successfully added the Google Maps Javascript API and the Google Places API Web Service. This finds nearby places based on the user's location (including suburbs, localities, businesses, etc) and appends them to the page. We'll need to tweak the parametres here to get the type of results that we're looking for.